compile

+++ tests/01-hello-world.cpp
@@ -45,7 +45,7 @@ static void test1() {
   p = new FloydShortestPaths<WeightedGraph<float>>(g);
   for (int i = 1; i <= 6; ++i) {
     for (int j = 1; j <= 6; ++j) {
-	  if (!p->HasPathOf(i, j)) printf("%f ", -1);
+	  if (!p->HasPathOf(i, j)) printf("%f ", (float)-1);
 	  else printf("%f ", p->TryGetDistanceOf(i, j).value());
     }
 	printf("\n");
